AssociationScript/src/transaction_processor.py
"""
Transaction Processor Module
Processes raw text to extract and structure transaction data
"""

import re
from datetime import datetime
from typing import List, Dict, Any
from dateutil.parser import parse as date_parse


class Transaction:
    """Class to represent a single transaction"""
    
    def __init__(self, date: str, description: str, debit: float = 0.0, 
                 credit: float = 0.0, balance: float = None):
        self.date = date
        self.description = description.strip()
        self.debit = debit
        self.credit = credit
        self.balance = balance
        self.transaction_type = "Credit" if credit > 0 else "Debit"
    
    def to_dict(self) -> Dict[str, Any]:
        """Convert transaction to dictionary for Excel export"""
        return {
            'Date': self.date,
            'Description': self.description,
            'Debit': self.debit if self.debit > 0 else '',
            'Credit': self.credit if self.credit > 0 else '',
            'Balance': self.balance if self.balance is not None else '',
            'Type': self.transaction_type
        }


class TransactionProcessor:
    """Class to process raw text and extract transactions"""
    
    def __init__(self):
        # Common date patterns (can be expanded based on your bank's format)
        self.date_patterns = [
            r'\d{2}/\d{2}/\d{4}',  # DD/MM/YYYY
            r'\d{4}-\d{2}-\d{2}',  # YYYY-MM-DD
            r'\d{2}-\d{2}-\d{4}',  # DD-MM-YYYY
            r'\d{2}\.\d{2}\.\d{4}', # DD.MM.YYYY
        ]
        
        # Amount patterns (with various formats)
        self.amount_patterns = [
            r'\d{1,3}(?:,\d{3})*\.\d{2}',  # 1,234.56
            r'\d+\.\d{2}',                 # 123.56
            r'\d{1,3}(?:,\d{3})*',         # 1,234
        ]
    
    def process_text(self, raw_text: str) -> List[Transaction]:
        """
        Process raw PDF text and extract transactions
        
        Args:
            raw_text (str): Raw text extracted from PDF
            
        Returns:
            List[Transaction]: List of processed transactions
        """
        print("🔍 Analyzing text patterns...")
        
        # Split text into lines for processing
        lines = raw_text.split('\n')
        transactions = []
        
        # Remove empty lines and clean up
        lines = [line.strip() for line in lines if line.strip()]
        
        print(f"📝 Processing {len(lines)} lines of text...")
        
        # For now, let's create a simple pattern-based approach
        # This will be enhanced as we test with real bank statements
        for i, line in enumerate(lines):
            transaction = self._extract_transaction_from_line(line)
            if transaction:
                transactions.append(transaction)
        
        print(f"💰 Found {len(transactions)} potential transactions")
        
        # If no transactions found, create sample data for testing
        if not transactions:
            print("ℹ️ No transactions detected. Creating sample data for testing...")
            transactions = self._create_sample_transactions()
        
        return transactions
    
    def _extract_transaction_from_line(self, line: str) -> Transaction:
        """
        Extract transaction data from a single line
        
        Args:
            line (str): Single line of text
            
        Returns:
            Transaction: Transaction object or None if not found
        """
        # Look for date pattern
        date_match = None
        for pattern in self.date_patterns:
            date_match = re.search(pattern, line)
            if date_match:
                break
        
        if not date_match:
            return None
        
        # Extract date
        date_str = date_match.group()
        
        # Look for amounts
        amounts = []
        for pattern in self.amount_patterns:
            amount_matches = re.findall(pattern, line)
            amounts.extend(amount_matches)
        
        if not amounts:
            return None
        
        # Clean and convert amounts
        cleaned_amounts = []
        for amount in amounts:
            cleaned = float(amount.replace(',', ''))
            cleaned_amounts.append(cleaned)
        
        # Extract description (everything except date and amounts)
        description = line
        description = re.sub(date_match.group(), '', description)
        for amount in amounts:
            description = description.replace(amount, '')
        description = description.strip()
        
        # Determine if it's credit or debit (this logic can be improved)
        # For now, assume last amount is the transaction amount
        if cleaned_amounts:
            amount = cleaned_amounts[-1]
            # Simple heuristic: if description contains certain keywords, it's likely a debit
            debit_keywords = ['ATM', 'WITHDRAWAL', 'PURCHASE', 'FEE', 'CHARGE']
            credit_keywords = ['DEPOSIT', 'CREDIT', 'SALARY', 'TRANSFER IN']
            
            is_debit = any(keyword in description.upper() for keyword in debit_keywords)
            is_credit = any(keyword in description.upper() for keyword in credit_keywords)
            
            if is_debit or (not is_credit):  # Default to debit if unclear
                return Transaction(date_str, description, debit=amount)
            else:
                return Transaction(date_str, description, credit=amount)
        
        return None
    
    def _create_sample_transactions(self) -> List[Transaction]:
        """Create sample transactions for testing purposes"""
        sample_transactions = [
            Transaction("01/01/2024", "Opening Balance", credit=1000.0, balance=1000.0),
            Transaction("02/01/2024", "ATM Withdrawal", debit=200.0, balance=800.0),
            Transaction("03/01/2024", "Salary Deposit", credit=2500.0, balance=3300.0),
            Transaction("04/01/2024", "Grocery Store Purchase", debit=150.75, balance=3149.25),
            Transaction("05/01/2024", "Online Transfer", debit=500.0, balance=2649.25),
            Transaction("06/01/2024", "Interest Credit", credit=25.50, balance=2674.75),
        ]
        return sample_transactions
