AssociationScript/requirements.txt
# Required packages for the Bank Statement to Excel Converter
pandas>=1.5.0
openpyxl>=3.1.0
pdfplumber>=0.9.0
python-dateutil>=2.8.2
xlsxwriter>=3.1.0
xlrd>=2.0.1
